Change threshold for zone determination to be 1/3 rather than 35. DUH! Why would any arbitrary number like 35 have been the threshold!?
This removes all the issues with missing points at the center of triangles EXCEPT for triangles with a total number of plants not divisible by 3, which still have a single missing point at their center. Still need to decide if this should be addressed in a more sophisticated manner than just throwing an error to prevent the user from making triangles with Ns that cause this.