Major updates to goelz functions to improve consistency and simplicity of function inputs and outputs. goelz() no longer as the reps and split arguments as those would never be used directly by a user, but are only used by goelz_optim(). Now, goelz_optim() does its own reformating to create the initial population. This ensures that an object of class goelz is always a tibble representing a single design and never a list of designs. This simplifies all other functions that take goelz objects as inputs.
goelz_optim() now also takes an object of class goelz to start it rather than the N argument. This is to maintain consistency with to be developed nelder_biculture_optim().

Also added select_optimal_goelz() for taking the output from goelz_optim() and parsing it to take the best design and outputing it as a goelz object.